+++ lib/src/eval/bindgen/context.dart
@@ -1,6 +1,8 @@
 class BindgenContext {
   final String uri;
   final Set<String> imports = {};
+  final Set<String> knownTypes = {};
+  final Set<String> unknownTypes = {};
   final bool wrap;
   final bool all;